[refactor] summary 관련 구조 리팩터링
