+++ src/main/java/com/cm2/entity/dto/summaryapi/PortMapping.java
@@ -1,4 +1,4 @@
-package com.cm2.entity.dto;
+package com.cm2.entity.dto.summaryapi;
 
 import lombok.Builder;
